add user tests and register tests
